Separated todo/user folders
Separated todo/user folders for convention.
