+++ server/src/main/java/umm3601/todo/Todo.java
@@ -1,4 +1,4 @@
-package umm3601.user;
+package umm3601.todo;
 
 public class Todo {
   public String _id;
