+++ server/src/main/java/umm3601/todo/TodoController.java
@@ -1,4 +1,4 @@
-package umm3601.user;
+package umm3601.todo;
 
 import io.javalin.http.Context;
 import io.javalin.http.NotFoundResponse;
